lushlyrics-webapp-django-main/requirements.txt
Django==3.0.7
youtube-search==1.1.1
spotipy==2.16.1
youtube-search==1.1.1
djangorestframework==3.11.0
django-allauth
dj-rest-auth